refactor: consolidate log contracts
