--- services/tracker/src/tracker/logs/__init__.py
@@ -1,21 +1,0 @@
-"""Provider-neutral benchmark log access."""
-
-from .base import (
-    LogEvent,
-    LogPage,
-    LogProvider,
-    LogProviderError,
-    RunLogReference,
-    RunTaskLogReference,
-    TaskLogReference,
-)
-
-__all__ = [
-    "LogEvent",
-    "LogPage",
-    "LogProvider",
-    "LogProviderError",
-    "RunLogReference",
-    "RunTaskLogReference",
-    "TaskLogReference",
-]
